web_resource_g2g/pyproject.toml
[tool.poetry]
name = "web-g2g"
version = "0.1.0"
description = ""
authors = ["Vladislav B <rolmf85@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
channels = "^4.2.0"
django-import-export = "^4.3.4"
aiogram = "^3.17.0"
coloredlogs = "^15.0.1"
mysqlclient = "^2.2.7"
whitenoise = "^6.9.0"
django = "^5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
